Move handleOnDateRangeSelect out of detailsHeader
