+++ src/routes/details/gcpDetails/detailsHeader.tsx
@@ -20,7 +20,6 @@ import type { ComputedGcpReportItemsParams } from 'routes/utils/computedReport/g
 import { getIdKeyForGroupBy } from 'routes/utils/computedReport/getComputedGcpReportItems';
 import { DateRangeType } from 'routes/utils/dateRange';
 import { filterProviders } from 'routes/utils/providers';
-import { getRouteForQuery } from 'routes/utils/query';
 import type { FetchStatus } from 'store/common';
 import { createMapStateToProps } from 'store/common';
 import { FeatureToggleSelectors } from 'store/featureToggle';
@@ -38,6 +37,7 @@ interface DetailsHeaderOwnProps {
   isCurrentMonthData?: boolean;
   isPreviousMonthData?: boolean;
   onCurrencySelect(value: string);
+  onDateRangeSelect(value: string);
   onGroupBySelect(value: string);
   query?: Query;
   report: GcpReport;
@@ -82,16 +82,13 @@ class DetailsHeaderBase extends React.Component<DetailsHeaderProps, any> {
   };
   public state: DetailsHeaderState = { ...this.defaultState };
 
-  private handleOnDateRangeSelected = (value: string) => {
-    const { query, router } = this.props;
+  private handleOnDateRangeSelect = (value: string) => {
+    const { onDateRangeSelect } = this.props;
 
     this.setState({ currentDateRangeType: value }, () => {
-      const newQuery = {
-        filter: {},
-        ...JSON.parse(JSON.stringify(query)),
-      };
-      newQuery.filter.time_scope_value = value === DateRangeType.previousMonth ? -2 : -1;
-      router.navigate(getRouteForQuery(newQuery, router.location, true), { replace: true });
+      if (onDateRangeSelect) {
+        onDateRangeSelect(value);
+      }
     });
   };
 
@@ -159,7 +156,7 @@ class DetailsHeaderBase extends React.Component<DetailsHeaderProps, any> {
                     isCurrentMonthData={isCurrentMonthData}
                     isDisabled={!showContent}
                     isPreviousMonthData={isPreviousMonthData}
-                    onSelect={this.handleOnDateRangeSelected}
+                    onSelect={this.handleOnDateRangeSelect}
                   />
                 </FlexItem>
               )}
