+++ src/routes/details/gcpDetails/gcpDetails.tsx
@@ -23,6 +23,7 @@ import { ProviderStatus } from 'routes/details/components/providerStatus';
 import { getIdKeyForGroupBy } from 'routes/utils/computedReport/getComputedGcpReportItems';
 import type { ComputedReportItem } from 'routes/utils/computedReport/getComputedReportItems';
 import { getUnsortedComputedReportItems } from 'routes/utils/computedReport/getComputedReportItems';
+import { DateRangeType } from 'routes/utils/dateRange';
 import { getGroupByTagKey } from 'routes/utils/groupBy';
 import { filterProviders, hasCurrentMonthData, hasPreviousMonthData } from 'routes/utils/providers';
 import { getRouteForQuery } from 'routes/utils/query';
@@ -281,6 +282,17 @@ class GcpDetails extends React.Component<GcpDetailsProps, GcpDetailsState> {
     }
   };
 
+  private handleOnDateRangeSelect = (value: string) => {
+    const { query, router } = this.props;
+
+    const newQuery = {
+      filter: {},
+      ...JSON.parse(JSON.stringify(query)),
+    };
+    newQuery.filter.time_scope_value = value === DateRangeType.previousMonth ? -2 : -1;
+    router.navigate(getRouteForQuery(newQuery, router.location, true), { replace: true });
+  };
+
   private handleOnExportModalClose = (isOpen: boolean) => {
     this.setState({ isExportModalOpen: isOpen });
   };
@@ -386,6 +398,7 @@ class GcpDetails extends React.Component<GcpDetailsProps, GcpDetailsState> {
           isCurrentMonthData={isCurrentMonthData}
           isPreviousMonthData={isPreviousMonthData}
           onCurrencySelect={() => handleOnCurrencySelect(query, router)}
+          onDateRangeSelect={this.handleOnDateRangeSelect}
           onGroupBySelect={this.handleOnGroupBySelect}
           query={query}
           report={report}
